+++ packages/lumx-react/src/components/input-helper/constants.ts
@@ -1,0 +1,13 @@
+const INPUT_HELPER_CONFIGURATION = {
+    error: {
+        color: 'red',
+    },
+    success: {
+        color: 'green',
+    },
+    warning: {
+        color: 'yellow',
+    },
+};
+
+export { INPUT_HELPER_CONFIGURATION };